BDOG-551 Move and refactor SlugInfoFlag
