+++ app/uk/gov/hmrc/cataloguefrontend/connector/ServiceDependenciesConnector.scala
@@ -19,6 +19,7 @@ package uk.gov.hmrc.cataloguefrontend.connector
 import javax.inject.{Inject, Singleton}
 import play.api.Logger
 import uk.gov.hmrc.cataloguefrontend.connector.model._
+import uk.gov.hmrc.cataloguefrontend.model.SlugInfoFlag
 import uk.gov.hmrc.cataloguefrontend.service.ServiceDependencies
 import uk.gov.hmrc.cataloguefrontend.util.UrlUtils.encodePathParam
 import uk.gov.hmrc.http.HeaderCarrier
@@ -27,22 +28,6 @@ import uk.gov.hmrc.play.bootstrap.http.HttpClient
 
 import scala.concurrent.{ExecutionContext, Future}
 import scala.util.control.NonFatal
-
-sealed trait SlugInfoFlag { def asString: String }
-object SlugInfoFlag {
-  case object Latest          extends SlugInfoFlag { val asString = "latest"        }
-  case object Production      extends SlugInfoFlag { val asString = "production"    }
-  case object ExternalTest    extends SlugInfoFlag { val asString = "external test" }
-  case object QA              extends SlugInfoFlag { val asString = "qa"            }
-  case object Staging         extends SlugInfoFlag { val asString = "staging"       }
-  case object Integration     extends SlugInfoFlag { val asString = "integration"   }
-  case object Dev             extends SlugInfoFlag { val asString = "development"   }
-
-  val values = List(Latest, Production, ExternalTest, QA, Staging, Integration, Dev)
-
-  def parse(s: String): Option[SlugInfoFlag] =
-    values.find(_.asString == s)
-}
 
 @Singleton
 class ServiceDependenciesConnector @Inject()(
