+++ app/views/BobbyRulesTrendPage.scala.html
@@ -16,8 +16,9 @@
 
 @import java.time.LocalDate
 @import play.twirl.api.Html
-@import uk.gov.hmrc.cataloguefrontend.connector.{SlugInfoFlag, Team}
+@import uk.gov.hmrc.cataloguefrontend.connector.Team
 @import uk.gov.hmrc.cataloguefrontend.connector.model.{BobbyRule, BobbyVersion, BobbyVersionRange, GroupArtefacts, ServiceWithDependency}
+@import uk.gov.hmrc.cataloguefrontend.model.SlugInfoFlag
 @import uk.gov.hmrc.cataloguefrontend.ViewMessages
 @import uk.gov.hmrc.cataloguefrontend.routes
 @import uk.gov.hmrc.cataloguefrontend.service.SearchByUrlService.FrontendRoutes
