+++ app/views/DependencyExplorerPage.scala.html
@@ -15,8 +15,9 @@
  *@
 
 @import play.api.mvc.Call
-@import uk.gov.hmrc.cataloguefrontend.connector.{SlugInfoFlag, Team}
+@import uk.gov.hmrc.cataloguefrontend.connector.Team
 @import uk.gov.hmrc.cataloguefrontend.connector.model.{BobbyVersion, BobbyVersionRange, GroupArtefacts, ServiceWithDependency, TeamName}
+@import uk.gov.hmrc.cataloguefrontend.model.SlugInfoFlag
 @import uk.gov.hmrc.cataloguefrontend.DependencyExplorerController.PieData
 @import uk.gov.hmrc.cataloguefrontend.ViewMessages
 @import uk.gov.hmrc.cataloguefrontend.routes
