+++ app/views/JdkAcrossEnvironmentsPage.scala.html
@@ -18,11 +18,10 @@
 @import uk.gov.hmrc.cataloguefrontend.service.ServiceDependencies
 @import uk.gov.hmrc.cataloguefrontend.service.ServiceDependency
 @import uk.gov.hmrc.cataloguefrontend.service.DependenciesService
-@import uk.gov.hmrc.cataloguefrontend.connector.model.{JDKVersion, Vendor, Oracle, OpenJDK}
-@import uk.gov.hmrc.cataloguefrontend.connector.SlugInfoFlag
+@import uk.gov.hmrc.cataloguefrontend.connector.model.{JDKUsageByEnv, JDKVersion, Vendor, Oracle, OpenJDK}
+@import uk.gov.hmrc.cataloguefrontend.model.SlugInfoFlag
 @import play.twirl.api.Html
 @import uk.gov.hmrc.cataloguefrontend.routes
-@import uk.gov.hmrc.cataloguefrontend.connector.model.JDKUsageByEnv
 
 
 @import uk.gov.hmrc.cataloguefrontend.{ routes => appRoutes }
