+++ app/views/JdkVersionPage.scala.html
@@ -19,7 +19,7 @@
 @import uk.gov.hmrc.cataloguefrontend.service.ServiceDependency
 @import uk.gov.hmrc.cataloguefrontend.service.DependenciesService
 @import uk.gov.hmrc.cataloguefrontend.connector.model.{JDKVersion, Oracle, OpenJDK}
-@import uk.gov.hmrc.cataloguefrontend.connector.SlugInfoFlag
+@import uk.gov.hmrc.cataloguefrontend.model.SlugInfoFlag
 @import play.twirl.api.Html
 @import uk.gov.hmrc.cataloguefrontend.routes
 
@@ -27,7 +27,8 @@
 
 @(jdkVersions: List[JDKVersion],
   flags:       Seq[SlugInfoFlag],
-  selected:    SlugInfoFlag)(implicit request: Request[_])
+  selected:    SlugInfoFlag
+  )(implicit request: Request[_])
 
 @standard_layout(s"JDK Versions", "dependencies") {
     <header>
