+++ app/views/ServiceInfoPage.scala.html
@@ -19,7 +19,6 @@
 @import play.twirl.api.Html
 @import uk.gov.hmrc.cataloguefrontend.connector.{DeploymentVO, RepositoryDetails}
 @import uk.gov.hmrc.cataloguefrontend.connector.model.{Dependencies, Dependency, Version}
-@import uk.gov.hmrc.cataloguefrontend.connector.SlugInfoFlag.Latest
 @import uk.gov.hmrc.cataloguefrontend.model.Environment
 @import uk.gov.hmrc.cataloguefrontend.service.RouteRulesService.{EnvironmentRoute, ServiceRoutes}
 @import uk.gov.hmrc.cataloguefrontend.shuttering.ShutterState
