+++ app/views/shuttering/shutterService/Page4.scala.html
@@ -17,7 +17,6 @@
 @import uk.gov.hmrc.cataloguefrontend.ViewMessages
 @import uk.gov.hmrc.cataloguefrontend.model.Environment
 @import uk.gov.hmrc.cataloguefrontend.shuttering.{ShutterStatusValue, ShutterWizardController, ShutterState, ShutterType}
-@import uk.gov.hmrc.cataloguefrontend.connector.SlugInfoFlag
 @import helper._
 @import partials.form_warnings
